Task 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,12 +9,12 @@ int main(void)
 	unsigned long int i;
 	unsigned long int t1 = 1;
 	unsigned long int t2 = 2;
-	int nextTerm = t1 + t2;
+	unsigned long int nextTerm = t1 + t2;
 	
-	printf("%d, %d", t1, t2);
+	printf("%lu, %lu", t1, t2);
 	for (i = 3; i <= 50; i++)
 	{
-		printf("%d", nextTerm);
+		printf("%lu", nextTerm);
 		t1 = t2;
 		t2 = nextTerm;
 		nextTerm = t1 + t2;
